Switched from getopt to clap
So getopt is an argument parser that is just as minimal as it is nice,
that is very. clap allows us to get a more extensive command line
argument repertoire and also lets us easily read dash-less arguemnts.

I will see how this goes and how much more needs to be patched...